# contracts/openapi/torgnexa-v1.yaml
openapi: 3.1.0
info:
  title: TORGNEXA API
  version: 0.16.0
servers:
  - url: /api/v1
paths:
...
  /inventory/warehouses:
    get:
      operationId: listInventoryWarehouses
      responses: {'200': {description: Canonical tenant warehouses}}
    post:
      operationId: createInventoryWarehouse
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WarehouseWrite'}}}}
      responses: {'201': {description: Warehouse created}, '409': {description: Duplicate code}}
...
  /inventory/fulfillment-allocations:
    get:
      operationId: listFulfillmentAllocations
      responses:
        '200': {description: Durable order-item reservation allocations and failover lineage}
    post:
      operationId: reserveFulfillmentAllocation
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FulfillmentAllocationWrite'}
      responses:
        '201': {description: Order item reserved at the requested allocatable warehouse}
        '400': {description: 'Invalid order item, warehouse, or insufficient ATP'}
        '409': {description: Idempotency or active-allocation conflict}
...
  /inventory/import:
    post:
      operationId: importInventory
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/InventoryImportWrite'}}}}
      responses: {'200': {description: Inventory rows imported through canonical audited mutations}, '400': {description: Invalid row or unknown SKU/warehouse}}
  /counterparties/search:
    get:
      operationId: searchLegalParties
      parameters:
        - {name: q, in: query, required: false, schema: {type: string, maxLength: 200}}
        - {name: inn, in: query, required: false, schema: {type: string, pattern: '^[0-9]{10,12}$'}}
        - {name: registration_id, in: query, required: false, schema: {type: string, maxLength: 32}}
        - {name: party_type, in: query, required: false, schema: {type: string, enum: [legal_entity, individual_entrepreneur, branch]}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Tenant-scoped canonical legal-party search results
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LegalPartySearchPage'}
        '400': {description: Invalid legal-party search query}
        '401': {description: Authenticated tenant scope missing}
...
  /inventory/positions:
    get:
      operationId: listInventoryPositions
      responses:
        '200':
          description: Tenant-scoped canonical inventory positions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 100
                    items:
                      type: object
                      required: [ID, SKU, ProductTitle, WarehouseCode, WarehouseName, OnHand, Reserved, Available, Unit, UpdatedAt]
                      properties:
                        ID: {type: string}
                        SKU: {type: string}
                        ProductTitle: {type: string}
                        WarehouseCode: {type: string}
                        WarehouseName: {type: string}
                        OnHand: {type: string}
                        Reserved: {type: string}
                        Available: {type: string}
                        Unit: {type: string}
                        UpdatedAt: {type: string, format: date-time}
        '401': {description: Authenticated tenant scope missing}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$'}
  schemas:
    WarehouseWrite:
      type: object
      properties:
        code: {type: string, minLength: 1, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 300}
        status: {type: string, enum: [active, disabled]}
        version: {type: integer, minimum: 1}
    FulfillmentAllocationWrite:
      type: object
      additionalProperties: false
      required: [order_item_id, warehouse_id]
      properties:
        order_item_id: {type: string, pattern: '^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$'}
        warehouse_id: {type: string, pattern: '^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$'}
...
    InventoryImportWrite:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: object
            required: [sku, warehouse_code, quantity, unit]
            properties:
              sku: {type: string, minLength: 1, maxLength: 128}
              warehouse_code: {type: string, minLength: 1, maxLength: 128}
              quantity: {type: string, pattern: '^(0|[1-9][0-9]*)([.][0-9]{1,9})?$'}
              unit: {type: string, pattern: '^[A-Z][A-Z0-9._-]{0,15}$'}
              reason: {type: string, maxLength: 128}
...
    LegalPartySearchPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            required: [party_type, party_id, code, display_name, status]
            properties:
              party_type: {type: string, enum: [legal_entity, individual_entrepreneur, branch]}
              party_id: {type: string}
              code: {type: string}
              display_name: {type: string}
              inn: {type: string}
              registration_id: {type: string}
              status: {type: string, enum: [draft, active, archived]}
...